+++ api/hydrophoneApi.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 	"log"
 	"net/http"
+	"net/url"
 	"os"
 	"runtime"
 	"strings"
@@ -281,6 +282,8 @@ func (a *Api) createAndSendNotification(req *http.Request, conf *models.Confirma
 	// Support address configuration contains the mailto we want to strip out
 	supportEmail := fmt.Sprintf("<a href=%s>%s</a>", a.Config.SupportURL, strings.Replace(a.Config.SupportURL, "mailto:", "", 1))
 
+	// Create an Encoded Email entry
+	content["EmailEncoded"] = url.QueryEscape(content["Email"].(string))
 	// Content collection is here to replace placeholders in template body/content
 	content["WebURL"] = a.getWebURL(req)
 	content["SupportURL"] = a.Config.SupportURL