+++ templates/preview/previewApi.go
@@ -6,6 +6,7 @@ import (
 	"io/ioutil"
 	"log"
 	"net/http"
+	"net/url"
 	"path"
 	"strings"
 
@@ -172,6 +173,7 @@ func (a *Api) generateEmail(templateName models.TemplateName, lang string) (stri
 	content := map[string]interface{}{
 		"Key":          "123456789123456789123456789123456789",
 		"Email":        "john@diabeloop.com",
+		"EncodedEmail": url.QueryEscape("john@diabeloop.com"),
 		"FullName":     "John Doe",
 		"PatientName":  "John Doe",
 		"WebPath":      "login",